refactor testing setup
