+++ src/util/testing/assets/setup-alice.json
@@ -1,0 +1,1 @@
+{"testUsers":{"alice":{"userName":"alice","keys":{"type":"MEMBER","name":"alice","generation":0,"signature":{"publicKey":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o=","secretKey":"+wvg9nNm7ZdZtZrp9iHLrjUdtIEnQ3ju01ovFgsG4gsGNbHoRdLUC5jSEB6GdoVqJxo/s0HYxPXELaEOBymHqg=="},"encryption":{"publicKey":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","secretKey":"UKQqQUjmgJhm4QDeex6j1lfQ/fwYyq8JEvGMV6JXbvE="},"secretKey":"f2aXBkFhRNZTFhEgOOOjxmIn6aDU/v9Hi95tK/8hwSHAoWvWphrj9G2wVz6nFeJt+LJY2oVRtH+L6bDA2bNMRA=="}}},"laptops":{"alice":{"deviceName":"laptop","userName":"alice","keys":{"type":"DEVICE","name":"alice::laptop","generation":0,"signature":{"publicKey":"D17qVUbuvctsDOZUev23MryIV9neUteYIEpHdXxtF68=","secretKey":"jZBd/jsSw1gxIr55TF4BYRVWRhWJqj8AWcfY633J0s0PXupVRu69y2wM5lR6/bcyvIhX2d5S15ggSkd1fG0Xrw=="},"encryption":{"publicKey":"/1/gr9b7wZxihhOrGUs3iHwVH64k6vWmmRjBCeDnTTE=","secretKey":"EoHtmOZphsICKvz9P6AF3GU++7mLH/IEd8hxUV86RUw="},"secretKey":"ofzWfbes92SdpivWyyQ0gsuHduyghOR28F2IT36ojO6tXPpkZhgY31f1vjctWCY6dVxJc9aT3702ZS8c2U5/tg=="}}},"phones":{"alice":{"deviceName":"phone","userName":"alice","keys":{"type":"DEVICE","name":"alice::phone","generation":0,"signature":{"publicKey":"Tjm8jW1gMM/G0IjWZRedewjedcwOsYS/X5CizkbBrGE=","secretKey":"CCG25PhW6NiGjeI+gSnbPvXVFlkib6jA/RAmcbPOGqhOObyNbWAwz8bQiNZlF517CN51zA6xhL9fkKLORsGsYQ=="},"encryption":{"publicKey":"O476+Uw7nL+MDFGQ0u0gaFzZetdcBVxabUAx8vmcNV8=","secretKey":"DO4DWCQoKSskYgGbXXOmOOW7rN38dkJW+74Y3ruVbXg="},"secretKey":"kEJlws9kJuja9Rk+l6PwGMl/ne/d8InZ8tN/R5kk9Plgzj935WmzoQVc9kFV/UBog9yrJgSGNjhtlmhxutRqrQ=="}}},"chain":{"root":"oxtiAmWfG8SDWTDWL1s8Yded288k2aJ54whTcQM46+10fSOsIWiss8cRl2VQhRUW/XQ4B8xyU9eGEMrLPJevvQ==","head":"oxtiAmWfG8SDWTDWL1s8Yded288k2aJ54whTcQM46+10fSOsIWiss8cRl2VQhRUW/XQ4B8xyU9eGEMrLPJevvQ==","links":{"oxtiAmWfG8SDWTDWL1s8Yded288k2aJ54whTcQM46+10fSOsIWiss8cRl2VQhRUW/XQ4B8xyU9eGEMrLPJevvQ==":{"body":{"type":"ROOT","payload":{"teamName":"Spies Я Us","rootMember":{"userName":"alice","keys":{"type":"MEMBER","name":"alice","generation":0,"encryption":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","signature":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o="},"roles":[]},"rootDevice":{"userName":"alice","deviceName":"laptop","keys":{"type":"DEVICE","name":"alice::laptop","generation":0,"encryption":"/1/gr9b7wZxihhOrGUs3iHwVH64k6vWmmRjBCeDnTTE=","signature":"D17qVUbuvctsDOZUev23MryIV9neUteYIEpHdXxtF68="}},"lockboxes":[{"encryptionKey":{"type":"EPHEMERAL","name":"EPHEMERAL","publicKey":"MHR6S1UmdKLFDRuL/vZAbVgdN/c5za9IEEjl/RAErAY="},"recipient":{"type":"MEMBER","name":"alice","generation":0,"encryption":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","signature":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o=","publicKey":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y="},"contents":{"type":"TEAM","name":"TEAM","generation":0,"encryption":"16ofOBVYHRxY8tGKqSvxiDd3AMD2mZ6XHF3QCsq7ZQY=","signature":"sgvB0GuHLeenfI0iK5ebQpgIWLyFKPy7AQBs4BBHz1I=","publicKey":"16ofOBVYHRxY8tGKqSvxiDd3AMD2mZ6XHF3QCsq7ZQY="},"encryptedPayload":"g6Vub25jZccYEjUkxHGo0eWzyld9uZ8b929ce0MrdGLFMqdtZXNzYWdlyAHYEml+76x9fgRyzcLGjOhzqmrAzYNEvTov30xaHUFLs36/bK/3StzC59I4J3ar7ZNw2qXVhb+yHGfnpnBBelgdHnMvDR6SNQ9EJ1vB6oYTF5kTZVene51FonUNM9lbwV0Jhyn26zeIYc7l9u1r5rGtEA51jGU4CjpFKblP8qzlJSM9KxKcdGoz8jUFmRc54EBIIj57vCm7Vn2w+quOZ6P+eOZarHvicfPyjWvMeBt+iTQ4an1F1E47hWB9//VZonhriMr0sITdUtgaude5NijBFOqPr/5ZaULcVp3enjzkVkyVV951LxZsWWiCt00GpSIA8v7+28cyQRnVjhgOgH5JrFQ4yRZuW7777yLo+7i7u8BMrPH0UZPNvRGtSkvop8hfj9j+bks99qAzXxP17yjTQ1H2hTZWALguZSb1KjRGxJtz/EGT5J0zORBNtT2VuE3rfm/LsuCOYUZzF9r7M1z6uMfzMUiYkdumlol87UZc32b84vw+VI26wPfXMr76/fglbpsSYdE8QgxVlTdFoNWYIf2WgQPGMTjQRrbpInzbKRyuxV0UUy0C5bqhKY7XZg8UHkTM1GaH/SHjr9HtXVc2Qt6zuvm5nLT+3gqWwfPLQt2RiztfglpFqsSvc2VuZGVyUHVibGljS2V5wA=="},{"encryptionKey":{"type":"EPHEMERAL","name":"EPHEMERAL","publicKey":"P5qxjOwxyM4qiiuy1IA6+HCSuXe67LP42k4HBeUqC1Y="},"recipient":{"type":"MEMBER","name":"alice","generation":0,"encryption":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","signature":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o=","publicKey":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y="},"contents":{"type":"ROLE","name":"admin","generation":0,"encryption":"Y9ncGlJCI8ySWyasEfFcrFs1pmB4DWuwiDkqTiSUK3g=","signature":"Elbs5h4SOyrTjYyqA1DqHEjeHh4Z/TuyarSLzBzXdII=","publicKey":"Y9ncGlJCI8ySWyasEfFcrFs1pmB4DWuwiDkqTiSUK3g="},"encryptedPayload":"g6Vub25jZccYEiL/kvZu839OxucYJ47fQ96OoQqMapyAw6dtZXNzYWdlyAHZEogZjauBoRDwpBWRM6zp4pB3Yv0LeecyXd8Y7qWBz2MbpBNNCm8Sob1VJ5+Y31UNnnOn9bkzd0bTLhKNp1bk0T3KW75DkG+eR7XQdiqqvRquIjIYCuuvuLjGTHtUSjoLhykXOWWzAAJyKdDG8+vrk7DfbpClgf0Y5baV1yqO0hKYPa/sNLQ11mtykToZQVucmUOgVr6sU/1H/E25sak/8jtnkshAReo6ndYpZU8td24foonhBx2Icf+xK6PUOaRA8Fg/MsBKXtF8YEH7pZLJ8Ml5j8GfsDY2IRgB9abjaCq/F3ynPzb7GNTqatgZZIvI0WV1aonkd6mVEXh9kHY0AO7kRGO1iabASNgP7iGbSQO3NtN8aEwCMdi6CijoH7XJ/s1JKGSxias0RjpkvN61CTvo4Lnsx/sDuzMRC2F9Qb27LokQU/e2ILmPItHQ8tQmuGZIIS3HYUuSQ69UV+C7h/VwY5SgFg3At0z062DSPr/2B8ZlgUSra0rNzZ+mFH/j3ur99rA0L9oPIk3ivnqIgDecr5CqIdbtHxKT1ERaFSPdLXmleFtyeqRpSTvQFSgImXRPdm1mOer5mOkQj6mvqvVLFAmiTIHjWRg2Rwwt35kGu9z2Oe0g//prr3NlbmRlclB1YmxpY0tlecA="},{"encryptionKey":{"type":"EPHEMERAL","name":"EPHEMERAL","publicKey":"mdVdGqSnUtd6FWQxkyhYgaq2IbVZerKQ1KXnlFOe+Bk="},"recipient":{"type":"DEVICE","name":"alice::laptop","generation":0,"encryption":"/1/gr9b7wZxihhOrGUs3iHwVH64k6vWmmRjBCeDnTTE=","signature":"D17qVUbuvctsDOZUev23MryIV9neUteYIEpHdXxtF68=","publicKey":"/1/gr9b7wZxihhOrGUs3iHwVH64k6vWmmRjBCeDnTTE="},"contents":{"type":"MEMBER","name":"alice","generation":0,"encryption":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","signature":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o=","publicKey":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y="},"encryptedPayload":"g6Vub25jZccYEnnbu7kWtdnUtXXg8b3xr1oh6N41N+2pb6dtZXNzYWdlyAHbEuxxSQ4cnnSn4DYaLjAlrq/gWw7ykxCnSFwUGgrqzZUEPnFzLqnxekdclxm0r0tU/h154aT4UBzMxiWNKSgofHvNxPICY/VtG+zz8W8hUyVjGRAW0nqQFJWLXh+RZSCnqpsBZm6VQxQhQ9XG9aQjchVgN8EuaSbKqWb7jAdcYfeOxmEfphb8L+moBrD56s66JO5h3hrjz3aP8n/U9fnUI35F9OoNK/jySPy0wHHAg/bWbvcHfkveJjNwFwi+LmkD6WsS13Z9f69bMGyQgZXgVibiBNb9BEp58Rh2noTshvFi6o2hVo+r6wIb1UX5W2Fg5yu0MWW6cHEzQmnh0ZnL1mftdc6FO1z6h7saM8wi+uNYhKLv5JDreEedzLPF7w0PEP1T7TDvaG4I16v+eraQdcXm9kFGaHScKcjFOcfcx0x1wb6jjaNjahbC0EL7J7/AoLcCSETCPhuhZIrsOdi7t93f9cbUvPjdCaAkJOChmGiTXjdsz/UFgsUkT1iOuTVVMrphWLssErsO3f8UAPRDQl8N8aUUOyjbKNMm23GTYWv/tvLxPl61qcIq/akDgZWJg44xVi1dOAPWb1b7oQRzo/Gy6kZzGGtJ+3qBs5o4q0Ak3Ig2wq/q0KI4usWvc2VuZGVyUHVibGljS2V5wA=="}]},"context":{"member":{"userName":"alice","keys":{"type":"MEMBER","name":"alice","generation":0,"encryption":"mFIpDwt2EptPHiB/n2o9JBEite14D0XtBws/1oTUf2Y=","signature":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o="},"roles":[]},"device":{"userName":"alice","deviceName":"laptop","keys":{"type":"DEVICE","name":"alice::laptop","generation":0,"encryption":"/1/gr9b7wZxihhOrGUs3iHwVH64k6vWmmRjBCeDnTTE=","signature":"D17qVUbuvctsDOZUev23MryIV9neUteYIEpHdXxtF68="}}},"timestamp":1612529560361,"prev":null},"hash":"oxtiAmWfG8SDWTDWL1s8Yded288k2aJ54whTcQM46+10fSOsIWiss8cRl2VQhRUW/XQ4B8xyU9eGEMrLPJevvQ==","signed":{"userName":"alice","signature":"gh6V9TEIkIbhz4p6h9TgThEyi17bC5CNON1SxbOJL33yAyI/a9laLD5ThD26ijvOPt/Zy5NMa/W+YW/5rTtZCQ==","key":"BjWx6EXS1AuY0hAehnaFaicaP7NB2MT1xC2hDgcph6o="}}}}}
