+++ src/util/testing/index.ts
@@ -1,3 +1,2 @@
 ﻿export * from '/util/testing/setup'
-export * from '/util/testing/storage'
-export * from '/util/testing/TestChannel'
+export * from '/util/testing/connectionHelpers'
