+++ src/util/testing/setup.ts
@@ -1,17 +1,14 @@
-﻿import memoize from 'fast-memoize'
-import fs from 'fs'
-import path from 'path'
+﻿import { cache } from './cache'
 import { InitialContext } from '/connection'
 import { Connection } from '/connection/Connection'
 import { LocalUserContext } from '/context'
-import { DeviceInfo, DeviceWithSecrets, getDeviceId } from '/device'
-import * as keysets from '/keyset'
+import * as devices from '/device'
+import { DeviceWithSecrets } from '/device'
 import { KeyType } from '/keyset'
 import { ADMIN } from '/role'
 import * as teams from '/team'
 import { Team } from '/team'
 import * as users from '/user'
-import * as devices from '/device'
 import { User } from '/user'
 import { arrayToMap, assert } from '/util'
 
@@ -35,78 +32,83 @@ export const setup = (_config: (TestUserSettings | string)[] = []) => {
   // Get a list of just user names
   const userNames = config.map((user) => user.user)
 
-  // Create users
-  const testUsers: Record<string, User> = userNames
-    .map((userName: string) =>
-      retrieveAsset(userName, () => {
-        const seed = userName // make these predictable
-        return users.create(userName, seed)
+  const cacheKey = 'setup-' + JSON.stringify(config)
+  const { testUsers, laptops, phones, chain } = cache(cacheKey, () => {
+    // Create users
+    const testUsers: Record<string, User> = userNames
+      .map((userName: string) => {
+        const randomSeed = userName // make these predictable
+        return users.create(userName, randomSeed)
       })
-    )
-    .reduce(arrayToMap('userName'), {})
+      .reduce(arrayToMap('userName'), {})
 
-  const testLaptops: Record<string, DeviceWithSecrets> = userNames
-    .map((userName: string) =>
-      retrieveAsset(`${userName}-laptop`, () => {
-        const deviceName = `${userName}'s laptop`
-        const seed = deviceName
-        const deviceInfo: DeviceInfo = { userName, deviceName }
-        return devices.create(deviceInfo, seed)
-      })
-    )
-    .reduce(arrayToMap('userName'), {})
+    const makeDevice = (userName: string, deviceName: string) => {
+      const key = `${userName}-${deviceName}`
+      const randomSeed = key
+      const device = devices.create({ deviceName, userName }, randomSeed)
+      return device
+    }
 
-  // Create team
-  const teamCacheKey = fileSystemSafe('team-' + JSON.stringify(config))
-  const chain = retrieveAsset(teamCacheKey, () => {
+    const laptops: Record<string, DeviceWithSecrets> = userNames
+      .map((userName: string) => makeDevice(userName, 'laptop'))
+      .reduce(arrayToMap('userName'), {})
+
+    const phones: Record<string, DeviceWithSecrets> = userNames
+      .map((userName: string) => makeDevice(userName, 'phone'))
+      .reduce(arrayToMap('userName'), {})
+
+    // Create team
     const founder = userNames[0] // e.g. alice
-    const founderContext = { user: testUsers[founder], device: testLaptops[founder] }
-    const teamSeed = 'seed123'
-    const team = teams.create('Spies Я Us', founderContext, teamSeed)
+    const founderContext = { user: testUsers[founder], device: laptops[founder] }
+    const teamName = 'Spies Я Us'
+    const randomSeed = teamName
+    const team = teams.create(teamName, founderContext, randomSeed)
+
     // Add members
     for (const { user: userName, admin = true, member = true } of config) {
       if (member && !team.has(userName)) {
         const user = testUsers[userName]
         const roles = admin ? [ADMIN] : []
-        const device = devices.redactDevice(testLaptops[userName])
+        const device = devices.redactDevice(laptops[userName])
         team.add(user, roles, device)
       }
     }
-    return team.chain
+    const chain = team.chain
+
+    return { testUsers, laptops, phones, chain }
   })
 
-  const makeUserStuff = ({ user: userName, member = true }: TestUserSettings) => {
+  const makeUserStuff = ({ user: userName, member = true }: TestUserSettings): UserStuff => {
     const user = testUsers[userName]
-    const device = testLaptops[userName]
+    const randomSeed = userName
+    const device = laptops[userName]
+    const phone = phones[userName]
     const localContext = { user, device }
     const team = member
       ? teams.load(chain, localContext) // members get a copy of the source team
-      : teams.create(userName, localContext) // non-members get a dummy empty placeholder team
+      : teams.create(userName, localContext, randomSeed) // non-members get a dummy empty placeholder team
+    const connectionContext = (member
+      ? { user, device, team }
+      : {
+          user,
+          device,
+          invitee: { type: KeyType.MEMBER, name: userName },
+          invitationSeed: '',
+        }) as InitialContext
+    const connection = {} as Record<string, Connection>
+    const getState = (peer: string) => connection[peer].state
 
-    const makeDevice = (deviceName: string) => {
-      const key = `${userName}-${deviceName}`
-      const device = retrieveAsset(key, () => {
-        return devices.create({ deviceName, userName }, key)
-      })
-
-      return device
-    }
-
-    const userStuff = {
+    return {
       userName,
       user,
       team,
       device,
       localContext,
-      phone: makeDevice('phone'),
-      connectionContext: member
-        ? { team, user, device }
-        : { invitee: { type: KeyType.MEMBER, name: userName }, invitationSeed: '', device },
-      connection: {} as Record<string, Connection>,
-      getState: (peer: string) => userStuff.connection[peer].state,
-    } as UserStuff
-
-    return userStuff
+      phone,
+      connectionContext,
+      connection,
+      getState,
+    }
   }
 
   const testUserStuff: Record<string, UserStuff> = config
@@ -116,139 +118,6 @@ export const setup = (_config: (TestUserSettings | string)[] = []) => {
   return testUserStuff
 }
 
-// HELPERS
-
-export const tryToConnect = async (a: UserStuff, b: UserStuff) => {
-  const aConnection = (a.connection[b.userName] = new Connection(a.connectionContext).start())
-  const bConnection = (b.connection[a.userName] = new Connection(b.connectionContext).start())
-
-  aConnection.pipe(bConnection).pipe(aConnection)
-}
-
-/** Connects the two members and waits for them to be connected */
-export const connect = async (a: UserStuff, b: UserStuff) => {
-  tryToConnect(a, b)
-  return connection(a, b)
-}
-
-/** Connects a (a member) with b (invited using the given seed). */
-export const connectWithInvitation = async (a: UserStuff, b: UserStuff, seed: string) => {
-  b.connectionContext = {
-    user: b.user,
-    device: b.device,
-    invitee: { type: KeyType.MEMBER, name: b.userName },
-    invitationSeed: seed,
-  }
-  return connect(a, b).then(() => {
-    // The connection now has the team object, so let's update our user stuff
-    b.team = b.connection[a.userName].team!
-  })
-}
-
-export const connectPhoneWithInvitation = async (a: UserStuff, seed: string) => {
-  const phoneContext = {
-    device: a.phone,
-    invitee: { type: KeyType.DEVICE, name: getDeviceId(a.phone) },
-    invitationSeed: seed,
-  } as InitialContext
-
-  const laptop = new Connection(a.connectionContext).start()
-  const phone = new Connection(phoneContext).start()
-
-  laptop.pipe(phone).pipe(laptop)
-
-  await all([laptop, phone], 'connected').then(() => {
-    a.team = laptop.team!
-  })
-}
-
-/** Passes if each of the given members is on the team, and knows every other member on the team */
-export const expectEveryoneToKnowEveryone = (...members: UserStuff[]) => {
-  for (const a of members)
-    for (const b of members) //
-      expect(a.team.has(b.userName)).toBe(true)
-}
-
-/** Disconnects the two members and waits for them to be disconnected */
-export const disconnect = (a: UserStuff, b: UserStuff) =>
-  Promise.all([
-    disconnection(a, b),
-    a.connection[b.userName].stop(),
-    b.connection[a.userName].stop(),
-  ])
-
-// PROMISIFIED EVENTS
-
-export const connection = async (a: UserStuff, b: UserStuff) => {
-  const connections = [a.connection[b.userName], b.connection[a.userName]]
-
-  // ✅ They're both connected
-  await all(connections, 'connected')
-
-  const sharedKey = connections[0].sessionKey
-  connections.forEach((connection) => {
-    expect(connection.state).toEqual('connected')
-    // ✅ They've converged on a shared secret key
-    expect(connection.sessionKey).toEqual(sharedKey)
-  })
-}
-
-export const updated = (a: UserStuff, b: UserStuff) => {
-  const connections = [a.connection[b.userName], b.connection[a.userName]]
-  return all(connections, 'updated')
-}
-
-export const disconnection = async (a: UserStuff, b: UserStuff, message?: string) => {
-  const connections = [a.connection[b.userName], b.connection[a.userName]]
-  const activeConnections = connections.filter((c) => c.state !== 'disconnected')
-
-  // ✅ They're both disconnected
-  await all(activeConnections, 'disconnected')
-
-  activeConnections.forEach((connection) => {
-    expect(connection.state).toEqual('disconnected')
-    // ✅ If we're checking for a message, it matches
-    if (message !== undefined) expect(connection.error!.message).toContain(message)
-  })
-}
-
-export const all = (connections: Connection[], event: string) =>
-  Promise.all(
-    connections.map((connection) => {
-      if (event === 'disconnect' && connection.state === 'disconnected') return true
-      if (event === 'connected' && connection.state === 'connected') return true
-      else return new Promise((resolve) => connection.on(event, () => resolve(true)))
-    })
-  )
-
-// TEST ASSETS
-
-const parseAssetFile = memoize((fileName: string) =>
-  JSON.parse(fs.readFileSync(fileName).toString())
-)
-
-const retrieveAsset = <T>(fileName: string, generateAsset: () => T): T => {
-  return generateAsset()
-
-  // const filePath = path.join(__dirname, `./assets/${fileName}.json`)
-
-  // // return cached object from assets folder if it exists
-  // if (fs.existsSync(filePath)) return parseAssetFile(filePath) as T
-
-  // // otherwise generate the asset
-  // const result: any = generateAsset()
-  // fs.writeFileSync(filePath, JSON.stringify(result))
-  // return result as T
-}
-
-const fileSystemSafe = (s: string) =>
-  s
-    .replace(/user/gi, '')
-    .replace(/[^a-z0-9]+/gi, '-')
-    .replace(/^-/i, '')
-    .replace(/-$/i, '')
-    .toLowerCase()
-
 // TYPES
 
 type TestUserSettings = {
@@ -257,7 +126,7 @@ type TestUserSettings = {
   member?: boolean
 }
 
-interface UserStuff {
+export interface UserStuff {
   userName: string
   user: User
   team: Team
